Calendar now displaying scheduled meals, but a few bugs remain, also reorganized some files in the frontend
