+++ bandits-ui/src/App.jsx
@@ -17,7 +17,7 @@ import Nav from 'react-bootstrap/Nav';
 import Navbar from 'react-bootstrap/Navbar';
 import NavDropdown from 'react-bootstrap/NavDropdown';
 import "./App.css";
-import { CalendarPage } from "./components/MealPlan/CalendarPage.jsx";
+import { CalendarPage } from "./components/Calendar/CalendarPage.jsx";
 
 
 // console.log(authenticated);
