--- bandits-ui/src/components/MealPlan/BasicCal.jsx
@@ -1,64 +1,0 @@
-import React, { Fragment, useMemo } from 'react'
-import PropTypes from 'prop-types'
-import moment from 'moment'
-import {
-  Calendar,
-  Views,
-  DateLocalizer,
-  momentLocalizer,
-} from 'react-big-calendar'
-import events from './events'
-import * as dates from './dates'
-
-const mLocalizer = momentLocalizer(moment)
-
-const ColoredDateCellWrapper = ({ children }) =>
-  React.cloneElement(React.Children.only(children), {
-    style: {
-      backgroundColor: 'lightblue',
-    },
-  })
-
-/**
- * We are defaulting the localizer here because we are using this same
- * example on the main 'About' page in Storybook
- */
-export default function Basic({
-  localizer = mLocalizer,
-  showDemoLink = true,
-  ...props
-}) {
-  const { components, defaultDate, max } = useMemo(
-    () => ({
-      components: {
-        timeSlotWrapper: ColoredDateCellWrapper,
-      },
-      defaultDate: new Date(2024, 9, 14),
-      max: dates.add(dates.endOf(new Date(2015, 17, 1), 'day'), -1, 'hours'),
-
-    }),
-    []
-  )
-
-  // {showDemoLink ? <DemoLink fileName="basic" /> : null}
-
-  return (
-    <Fragment>
-      <div className="height600" {...props}>
-        <Calendar
-          components={components}
-          defaultDate={defaultDate}
-          events={events}
-          localizer={localizer}
-          max={max}
-          showMultiDayTimes
-          step={60}
-        />
-      </div>
-    </Fragment>
-  )
-}
-Basic.propTypes = {
-  localizer: PropTypes.instanceOf(DateLocalizer),
-  showDemoLink: PropTypes.bool,
-}
