+++ bandits-ui/src/services/mealService.js
@@ -1,6 +1,6 @@
 import axios from "axios";
 
-const LOCAL_API_BASE_URL = "http://localhost:8080";
+const LOCAL_API_MEAL_URL = "http://localhost:8080/api/meals";
 
 const RECIPE_SEARCH_BASE_URL = "https://api.edamam.com/api/recipes/v2?type=public";
 const RECIPE_SEARCH_APP_KEY = "f7b7c5856e28029b8e7ab08182ab9fe1"
@@ -19,7 +19,7 @@ const RECIPE_SEARCH_APP_ID = "9cf8e5c7";
 
 export const fetchMeals = async () => {
   try {
-    const response = await axios.get(`${LOCAL_API_BASE_URL}/api/meals/saved`, {withCredentials:true});
+    const response = await axios.get(`${LOCAL_API_MEAL_URL}/saved`, {withCredentials:true});
     console.log("response", response);
     return response.data;
   } catch (error) {
@@ -28,21 +28,22 @@ export const fetchMeals = async () => {
   }
 };
 
-export const fetchCalendarMeals = async () => {
-  try {
-    const response = await axios.get(`${LOCAL_API_BASE_URL}/api/calendar`, {withCredentials:true});
-    console.log("response.data, from service", response.data);
-    return response.data;
-  } catch (error) {
-    console.error("There was an error fetching the meals!", error);
-    throw error;
-  }
-};
+// // Moved to calendarService, left here in case something breaks...
+// export const fetchCalendarMeals = async () => {
+//   try {
+//     const response = await axios.get(`${LOCAL_API_MEAL_URL}/api/calendar`, {withCredentials:true});
+//     console.log("response.data, from service", response.data);
+//     return response.data;
+//   } catch (error) {
+//     console.error("There was an error fetching the meals!", error);
+//     throw error;
+//   }
+// };
 
 export const addMeal = async (uri, label, mealType, date) => {
   try {
     // console.log("uri", uri, "label", label, "image", image, "source", source);
-    await axios.post(`${LOCAL_API_BASE_URL}/api/meals/new`, null, {
+    await axios.post(`${LOCAL_API_MEAL_URL}/new`, null, {
       params: { uri, label, mealType, date },
       withCredentials: true,
     });
@@ -57,7 +58,7 @@ export const addMeal = async (uri, label, mealType, date) => {
 // export const addMeal = async (meal) => {
 //   try {
 //     // console.log("uri", uri, "label", label, "image", image, "source", source);
-//     const response = await axios.post(`${LOCAL_API_BASE_URL}/api/meals/new`, meal,
+//     const response = await axios.post(`${LOCAL_API_MEAL_URL}/api/meals/new`, meal,
 //       {
 //         headers: { 'content-type': 'application/json' },
 //         withCredentials: true,
@@ -73,7 +74,7 @@ export const addMeal = async (uri, label, mealType, date) => {
 
 export const deleteMeal = async (recipeId) => {
   try {
-    await axios.post(`${LOCAL_API_BASE_URL}/api/meals/delete`, null, {
+    await axios.post(`${LOCAL_API_MEAL_URL}/delete`, null, {
       params: { recipeId },
       withCredentials: true,
     });
